Crate iroh_net_report

Checks the network conditions from the current host.

NetReport is responsible for finding out the network conditions of the current host, like whether it is connected to the internet via IPv4 and/or IPv6, what the NAT situation is etc and reachability to the configured relays.

Structs

Addr

Sender to the main service.

Client

Client to run net_reports.

Metrics

Enum of metrics for the module

RelayLatencies

Latencies per relay node.

Report

A net_report report.

Functions

os_has_ipv6

Test if IPv6 works at all, or if it’s been hard disabled at the OS level.