iroh_net_report
Module task

Async rust task spawning and utilities that work natively (using tokio) and in browsers (using wasm-bindgen-futures).

Structs

AbortOnDropHandle

A wrapper around a tokio::task::JoinHandle, which aborts the task when it is dropped.

JoinError

Task failed to execute to completion.

JoinHandle

An owned permission to join on a task (await its termination).

JoinSet

A collection of tasks spawned on a Tokio runtime.

Functions

spawn

Spawns a new asynchronous task, returning a JoinHandle for it.