Crate iroh_net_report

Checks the network conditions from the current host.

NetReport is responsible for finding out the network conditions of the current host, like whether it is connected to the internet via IPv4 and/or IPv6, what the NAT situation is etc and reachability to the configured relays.

Structs

Addr

Sender to the main service.

Client

Client to run net_reports.

IpMappedAddr

A mirror for the NodeIdMappedAddr, mapping a fake Ipv6 address with an actual IP address.

IpMappedAddrError

Can occur when converting a SocketAddr to an IpMappedAddr

IpMappedAddrs

A Map of IpMappedAddrs to SocketAddr.

Metrics

Enum of metrics for the module

Options

Options for running probes

QuicConfig

Pieces needed to do QUIC address discovery.

RelayLatencies

Latencies per relay node.

Report

A net_report report.

Constants

MAPPED_ADDR_PORT

The dummy port used for all mapped addresses

Functions

bind_local_stun_socketstun-utils

Attempts to bind a local socket to send STUN packets from.

os_has_ipv6

Test if IPv6 works at all, or if it’s been hard disabled at the OS level.